requirements.txt
amqp==2.4.2
beautifulsoup4==4.7.1
billiard==3.5.0.5
bs4==0.0.1
celery==4.2.1
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
djangorestframework==3.9.2
idna==2.8
kombu==4.4.0
lxml==4.3.2
psycopg2==2.7.7
python-dateutil==2.8.0
pytz==2018.9
redis==3.2.0
requests==2.21.0
six==1.12.0
soupsieve==1.8
urllib3==1.24.1
vine==1.2.0
